Refactor into individual components
